requirements.txt
aioredis==1.3.1
amqp==5.1.1
ansible-runner==2.1.2
asgiref==3.6.0
assistant==0.9.5a0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
bcrypt==4.0.1
billiard==3.6.4.0
cachetools==5.2.1
celery==5.2.3
cffi==1.15.1
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
commonmark==0.9.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
daphne==3.0.2
Django==4.0.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-filter==22.1
django-redis==5.2.0
django-rest-framework-mongoengine==3.4.1
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
dmt==1.5.6
dnspython==2.3.0
docutils==0.19
drf-spectacular==0.21.2
exceptiongroup==1.1.0
filelock==3.9.0
Flask==2.2.2
ftfy==6.1.1
giteapy @ git+https://github.com/dblueai/giteapy.git@001e9b66795a6d34146c8532e9d8e648d5b93e59
google-auth==2.16.0
h11==0.14.0
halo==0.0.31
hiredis==2.1.1
huggingface-hub==0.11.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
kubernetes==25.3.0
lockfile==0.12.2
log-symbols==0.0.14
MarkupSafe==2.1.2
mongoengine==0.26.0
msgpack==1.0.4
mysqlclient==2.1.1
numpy==1.24.1
oauthlib==3.2.2
outcome==1.2.0
packaging==22.0
paramiko==3.0.0
pexpect==4.8.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
Pygments==2.14.0
PyJWT==2.6.0
pymongo==4.3.3
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-crontab==2.7.1
python-daemon==2.3.2
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0
PyYAML==6.0
questionary==1.10.0
redis==4.4.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.0.0
rsa==4.9
selenium==4.7.2
sentencepiece==0.1.97
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
spinners==0.0.24
sqlparse==0.4.3
termcolor==2.2.0
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.25.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.2.2
wsproto==1.2.0
xonsh==0.13.4
zope.interface==5.5.2
